requirements.txt
fastapi
uvicorn
neo4j
google-generativeai
pydantic
python-dotenv